requirements.txt
# Production dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development dependencies (for CI/CD)
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
requests>=2.31.0